dicom: split existing Resource classes into separate files
