+++ plugins/dicom/server/rest/studies.py
@@ -25,60 +25,9 @@ from girder.api.describe import Description
 from girder.api.rest import Resource, getUrlParts
 from girder.api import access
 
-__all__ = ['dicomStudies', 'dicomSeries', 'dicomInstances']
+from error_descriptions import _describe_wadors_errors, _describe_qidors_errors, _describe_stowrs_errors
+from param_descriptions import *
 
-# QIDO-RS
-SearchForDescription = "Searches for DICOM %s that match specified search parameters and returns a" \
-                       "list of matching %s and the requested attributes for each %s."
-
-QueryParamDescription = "{attributeID}={value}"
-FuzzyMatchingParamDescription = "Set to 'true' to perform additional fuzzy semantic matching of person names."
-LimitParamDescription = "If the limit query key is not specified or its value exceeds the total number" \
-                        " of matching results then {maximumResults} is the lesser of the number of matching" \
-                        "results and the maximum number of results supported by the Server."
-OffsetParamDescription = "If the offset query key is not specified or its value is less than zero then it defaults to zero."
-
-def _describe_wadors_errors(func):
-    return (func.description
-        .errorResponse('Partial Content: Accept type, Transfer Syntax or decompression method '
-                       'supported for some but not all requested content.', 206)
-        .errorResponse('Bad Request: Malformed resource.', 400)
-        .errorResponse('Not Found: Specified resource does not exist.', 404)
-        .errorResponse('Not Acceptable: Accept type, Transfer Syntax or decompression method not supported.', 406)
-        .errorResponse('Gone: Specified resource was deleted.', 410)
-        .errorResponse('Busy: Service is unavailable.', 503))
-
-def _describe_qidors_errors(func):
-    return (func.description
-        .errorResponse('Bad Request: The QIDO-RS Provider was unable to perform the query '
-                       'because the Service Provider cannot understand the query component.', 400)
-        .errorResponse('Unauthorized: The QIDO-RS Provider refused to perform the query because '
-                       'the client is not authenticated.', 401)
-        .errorResponse('Forbidden: The QIDO-RS Provider understood the request, but is refusing to '
-                       'perform the query (e.g., an authenticated user with insufficient privileges).', 403)
-        .errorResponse('Request entity too large: The query was too broad and a narrower query or '
-                       'paging should be requested. The use of this status code should be documented '
-                       'in the conformance statement.', 413)
-        .errorResponse('Busy: Service is unavailable.', 503))
-
-def _describe_stowrs_errors(func):
-    return (func.description
-            .errorResponse('Bad Request: This indicates that the STOW-RS Service was unable to store '
-                           'any instances due to bad syntax.', 400)
-            .errorResponse('Unauthorized: This indicates that the STOW-RS Service refused to create or '
-                           'append any instances because the client is not authorized.', 401)
-            .errorResponse('Forbidden: This indicates that the STOW-RS Service understood the request, '
-                           'but is refusing to fulfill it (e.g., an authorized user with insufficient '
-                           'privileges).', 403)
-            .errorResponse('Conflict: This indicates that the STOW-RS Service request was formed correctly '
-                           'but the service was unable to store any instances due to a conflict in the request '
-                           '(e.g., unsupported SOP Class or StudyInstanceUID mismatch). This may also be used to '
-                           'indicate that a STOW-RS Service was unable to store any instances for a mixture of '
-                           'reasons. Additional information regarding the instance errors can be found in the '
-                           'XML response message body.', 409)
-            .errorResponse('Unsupported Media Type: This indicates that the STOW-RS Service does not support '
-                           'the Content-Type specified in the storage request (e.g., the service does not '
-                           'support JSON metadata).)', 415))
 
 class dicomStudies(Resource):
 
@@ -349,68 +298,3 @@ class dicomStudies(Resource):
         Description(StoreInstancesDescription)
         .param('StudyInstanceUID', StudyInstanceUIDDescription, paramType='path'))
     storeInstancesInStudy.description = _describe_stowrs_errors(storeInstancesInStudy)
-
-
-class dicomSeries(Resource):
-    def __init__(self):
-        self.resourceName = 'series'
-
-        ###########
-        # QIDO-RS #
-        ###########
-
-        # {+SERVICE}/series{?query*,fuzzymatching,limit,offset}
-        self.route('GET', (), self.searchForSeries)
-
-    ###########
-    # QIDO-RS #
-    ###########
-
-    @access.user
-    def searchForSeries(self, params):
-        pass
-
-    searchForSeries.description = (
-        Description(SearchForDescription % ('Series', 'series', 'series'))
-        .param('query', QueryParamDescription,
-               required=False)
-        .param('fuzzymatching', FuzzyMatchingParamDescription,
-               required=False, dataType='boolean', default=False)
-        .param('limit', LimitParamDescription,
-               required=False, dataType='integer')
-        .param('offset', OffsetParamDescription,
-               required=False, dataType='integer', default=0))
-    searchForSeries.description = _describe_qidors_errors(searchForSeries)
-
-
-class dicomInstances(Resource):
-    def __init__(self):
-        self.resourceName = 'instances'
-
-        ###########
-        # QIDO-RS #
-        ###########
-
-        # {+SERVICE}/instances{?query*,fuzzymatching,limit,offset}
-        self.route('GET', (), self.searchForInstances)
-
-    ###########
-    # QIDO-RS #
-    ###########
-
-    @access.user
-    def searchForInstances(self, params):
-        pass
-
-    searchForInstances.description = (
-        Description(SearchForDescription % ('Instances', 'instances', 'instance'))
-        .param('query', QueryParamDescription,
-               required=False)
-        .param('fuzzymatching', FuzzyMatchingParamDescription,
-               required=False, dataType='boolean', default=False)
-        .param('limit', LimitParamDescription,
-               required=False, dataType='integer')
-        .param('offset', OffsetParamDescription,
-               required=False, dataType='integer', default=0))
-    searchForInstances.description = _describe_qidors_errors(searchForInstances)
-
